services/llama-agent/requirements.txt
llama-index==0.10.0
llama-index-llms-openai==0.1.0
pydantic==2.5.0
httpx==0.25.0
python-dotenv==1.0.0
structlog==23.2.0
uvloop==0.19.0